+++ src/commands/meme.js
@@ -9,13 +9,13 @@ module.exports = {
 	async execute(interaction) {
 		const data = await fetch('http://meme-api.herokuapp.com/gimme').then(res => res.json());
 
-		const embed = new MessageEmbed();
-		embed.setAuthor(data.author, 'https://www.redditstatic.com/avatars/defaults/v2/avatar_default_1.png', 'https://reddit.com/u/' + data.author);
-		embed.setTitle(data.title);
-		embed.setURL(data.postLink);
-		embed.setImage(data.url);
-		embed.setFooter(`r/${data.subreddit} • ${data.ups} 👍`);
-		embed.setTimestamp();
+		const embed = new MessageEmbed()
+			.setAuthor(data.author, 'https://www.redditstatic.com/avatars/defaults/v2/avatar_default_1.png', 'https://reddit.com/u/' + data.author)
+			.setTitle(data.title)
+			.setURL(data.postLink)
+			.setImage(data.url)
+			.setFooter(`r/${data.subreddit} • ${data.ups} 👍`)
+			.setTimestamp();
 
 		interaction.reply({ embeds: [embed] });
 	},