+++ src/commands/reddit.js
@@ -15,14 +15,14 @@ module.exports = {
 			return interaction.reply({ content: data.message, ephemeral: true });
 		}
 
-		const embed = new MessageEmbed();
-		embed.setColor('#FF4300');
-		embed.setAuthor(data.author, 'https://www.redditstatic.com/avatars/defaults/v2/avatar_default_1.png', 'https://reddit.com/u/' + data.author);
-		embed.setTitle(data.title);
-		embed.setURL(data.postLink);
-		embed.setImage(data.url);
-		embed.setFooter(`r/${data.subreddit} • ${data.ups} 👍`);
-		embed.setTimestamp();
+		const embed = new MessageEmbed()
+			.setColor('#FF4300')
+			.setAuthor(data.author, 'https://www.redditstatic.com/avatars/defaults/v2/avatar_default_1.png', 'https://reddit.com/u/' + data.author)
+			.setTitle(data.title)
+			.setURL(data.postLink)
+			.setImage(data.url)
+			.setFooter(`r/${data.subreddit} • ${data.ups} 👍`)
+			.setTimestamp();
 
 		interaction.reply({ embeds: [embed] });
 	},