+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Cleaner code for `reddit` & `meme`
+- Adding `hug` command (subcommand of `interact`)
+
 ## [1.3.1-alpha] - 2021-11-08
 ### Changed
 - Bump @discordjs/builders from 0.7.0 to 0.8.2 (#12)